Erase from fav list and comic detail from fav list.
Enabled delete button in fav list - comic adding a method that calls the delete comic from fav list service.
When the image is clicked in the fav-list comic the comic is passed to the comic.modal.service in order to be displayed. another method has been added to the service in order to receive individual comics. the modal comic component has also been updated to subscribe to this new service and to update the comic attribute, since no new http request is needed :D